Update aws-cs-s3-folder (tested locally - works)
