+++ .github/workflows/performance_metrics_cron.yml
@@ -84,7 +84,7 @@ jobs:
       fail-fast: false
       matrix:
         dotnet-version:
-          - 3.1.301
+          - 6.0.300
         go-version:
           - 1.16.x
         node-version:
