+++ .github/workflows/run-tests-command.yml
@@ -61,7 +61,7 @@ jobs:
       fail-fast: false
       matrix:
         dotnet-version:
-        - 3.1.x
+        - 6.0.x
         platform:
         - ubuntu-latest
         source-dir:
@@ -179,7 +179,7 @@ jobs:
       fail-fast: false
       matrix:
         dotnet-version:
-        - 3.1.301
+        - 6.0.300
         go-version:
         - 1.16.x
         node-version:
@@ -514,7 +514,7 @@ jobs:
       fail-fast: false
       matrix:
         dotnet-version:
-        - 3.1.301
+        - 6.0.300
         go-version:
         - 1.16.x
         node-version:
