+++ .github/workflows/smoke-test-cli-command.yml
@@ -39,7 +39,7 @@ jobs:
       fail-fast: false
       matrix:
         dotnet-version:
-        - 3.1.x
+        - 6.0.x
         platform:
         - ubuntu-latest
         source-dir:
@@ -371,7 +371,7 @@ jobs:
       fail-fast: false
       matrix:
         dotnet-version:
-        - 3.1.301
+        - 6.0.300
         go-version:
         - 1.16.x
         node-version:
@@ -452,7 +452,7 @@ jobs:
       fail-fast: false
       matrix:
         dotnet-version:
-        - 3.1.301
+        - 6.0.300
         go-version:
         - 1.16.x
         node-version:
