+++ .github/workflows/smoke-test-provider-command.yml
@@ -39,7 +39,7 @@ jobs:
       fail-fast: false
       matrix:
         dotnet-version:
-        - 3.1.x
+        - 6.0.x
         platform:
         - ubuntu-latest
         source-dir:
@@ -155,7 +155,7 @@ jobs:
       fail-fast: false
       matrix:
         dotnet-version:
-        - 3.1.301
+        - 6.0.300
         go-version:
         - 1.16.x
         node-version:
@@ -245,7 +245,7 @@ jobs:
       fail-fast: false
       matrix:
         dotnet-version:
-        - 3.1.301
+        - 6.0.300
         go-version:
         - 1.16.x
         languages:
@@ -364,7 +364,7 @@ jobs:
       fail-fast: false
       matrix:
         dotnet-version:
-        - 3.1.301
+        - 6.0.300
         go-version:
         - 1.16.x
         node-version:
@@ -445,7 +445,7 @@ jobs:
       fail-fast: false
       matrix:
         dotnet-version:
-        - 3.1.301
+        - 6.0.300
         go-version:
         - 1.16.x
         node-version:
